moved tailwind config to root; fixed linting error
